game-bot/requirements.txt
Appium-Python-Client==4.2.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
easyocr==1.7.2
exceptiongroup==1.2.2
filelock==3.16.1
fsspec==2024.9.0
h11==0.14.0
humanize==4.10.0
idna==3.10
imageio==2.35.1
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
numpy==2.0.2
opencv-python-headless==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
pendulum==3.0.0
pillow==10.4.0
pyclipper==1.3.0.post5
PySocks==1.7.1
python-bidi==0.6.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.0.8
scikit-image==0.24.0
scipy==1.13.1
selenium==4.25.0
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
structlog==24.4.0
sympy==1.13.3
tifffile==2024.8.30
time-machine==2.15.0
torch==2.4.1
torchvision==0.19.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
